requirements.txt
Flask==2.2.3
hnswlib==0.7.0
numpy==1.23.5
openai==0.27.2
openai_whisper==20230314
PyAudio==0.2.13
pyttsx3==2.90
readability_lxml==0.8.1
requests==2.28.2
selenium==4.8.3
sentence_transformers==2.2.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
webrtcvad==2.0.10
whisper==1.1.10
